report_api/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
numpy==1.24.3
pandas==2.0.3
mysql-connector-python==8.1.0
psycopg2-binary==2.9.7
google-generativeai>=0.5.2